Remove the legacy agent tool RPC flow
