+++ packages/ballerina-core/src/flow-model/flow-model-utils.ts
@@ -16,10 +16,9 @@
  * under the License.
  */
 
-import { AgentToolData, FlowNode, Property } from "../interfaces/bi";
+import { FlowNode } from "../interfaces/bi";
 
 import { Flow } from "../interfaces/bi";
-import { ToolParameters } from "../rpc-types/ai-agent/interfaces";
 import { BaseVisitor } from "./BaseVisitor";
 
 const metaNodes = ["viewState", "position", "parent"];
@@ -93,43 +92,4 @@ export function traverseNode(node: FlowNode, visitor: BaseVisitor, parent?: Flow
     if (endVisitFn) {
         endVisitFn.bind(visitor)(node, parent);
     }
-}
-
-export function buildAgentToolNode(wrappedNode: FlowNode, toolName: string, description: string, connection: string,
-    toolParameters?: ToolParameters): FlowNode {
-    const auth = wrappedNode.codedata.data?.auth;
-    const data: AgentToolData = {
-        node: wrappedNode,
-        connection,
-        description,
-        ...(typeof auth === "string" ? { auth } : {}),
-    };
-    return createAgentToolNode(toolName, data, toolParameters ? { parameters: toolParameters } : {});
-}
-
-export function buildAgentCallToolNode(toolName: string, agentVarName: string, includeContext: boolean,
-    description: string): FlowNode {
-    const data: AgentToolData = { toolKind: "AGENT_CALL", agentVarName, includeContext, description };
-    return createAgentToolNode(toolName, data);
-}
-
-function createAgentToolNode(toolName: string, data: AgentToolData, extraProperties: FlowNode["properties"] = {}): FlowNode {
-    return {
-        id: "0",
-        metadata: { label: "Agent Tool", description: "" },
-        codedata: { node: "AGENT_TOOL", isNew: true, data },
-        properties: {
-            functionName: {
-                metadata: { label: "Name", description: "Name of the tool" },
-                valueType: "IDENTIFIER",
-                value: toolName,
-                optional: false,
-                editable: true,
-                advanced: false,
-            } as Property,
-            ...extraProperties,
-        },
-        branches: [],
-        returning: false,
-    };
 }
