+++ packages/ballerina-extension/src/rpc-managers/ai-agent/rpc-handler.ts
@@ -19,14 +19,11 @@
  */
 import {
     AiModuleOrgRequest,
-    AIAgentRequest,
-    AIAgentToolsUpdateRequest,
     AIModelsRequest,
     AINodesRequest,
     AIToolRequest,
     AIToolsRequest,
     configureDefaultModelProvider,
-    createAIAgent,
     DefaultProviderKind,
     getAiModuleOrg,
     getAllAgents,
@@ -39,7 +36,6 @@ import {
     McpToolsRequest,
     McpToolUpdateRequest,
     MemoryManagersRequest,
-    updateAIAgentTools,
     updateMCPToolKit,
     AIGetPackageVersionRequest,
     getPackageVersion,
@@ -61,7 +57,5 @@ export function registerAiAgentRpcHandlers(messenger: Messenger) {
     messenger.onRequest(fixMissingImports, () => rpcManger.fixMissingImports());
     messenger.onRequest(getPackageVersion, (args: AIGetPackageVersionRequest) => rpcManger.getPackageVersion(args));
     messenger.onNotification(configureDefaultModelProvider, (kind: DefaultProviderKind) => rpcManger.configureDefaultModelProvider(kind));
-    messenger.onRequest(createAIAgent, (args: AIAgentRequest) => rpcManger.createAIAgent(args));
-    messenger.onRequest(updateAIAgentTools, (args: AIAgentToolsUpdateRequest) => rpcManger.updateAIAgentTools(args));
     messenger.onNotification(updateMCPToolKit, (args: McpToolUpdateRequest) => rpcManger.updateMCPToolKit(args));
 }
