+++ packages/ballerina-rpc-client/src/rpc-clients/ai-agent/rpc-client.ts
@@ -21,9 +21,6 @@ import {
     AIAgentAPI,
     AiModuleOrgRequest,
     AiModuleOrgResponse,
-    AIAgentRequest,
-    AIAgentResponse,
-    AIAgentToolsUpdateRequest,
     AIModelsRequest,
     AIModelsResponse,
     AINodesRequest,
@@ -33,7 +30,6 @@ import {
     AIToolsRequest,
     AIToolsResponse,
     configureDefaultModelProvider,
-    createAIAgent,
     DefaultProviderKind,
     getAiModuleOrg,
     getAllAgents,
@@ -48,7 +44,6 @@ import {
     McpToolUpdateRequest,
     MemoryManagersRequest,
     MemoryManagersResponse,
-    updateAIAgentTools,
     updateMCPToolKit,
     getPackageVersion,
     fixMissingImports,
@@ -108,14 +103,6 @@ export class AiAgentRpcClient implements AIAgentAPI {
     configureDefaultModelProvider(kind: DefaultProviderKind): Promise<void> {
         this._messenger.sendNotification(configureDefaultModelProvider, HOST_EXTENSION, kind);
         return Promise.resolve();
-    }
-
-    createAIAgent(params: AIAgentRequest): Promise<AIAgentResponse> {
-        return this._messenger.sendRequest(createAIAgent, HOST_EXTENSION, params);
-    }
-
-    updateAIAgentTools(params: AIAgentToolsUpdateRequest): Promise<AIAgentResponse> {
-        return this._messenger.sendRequest(updateAIAgentTools, HOST_EXTENSION, params);
     }
 
     updateMCPToolKit(params: McpToolUpdateRequest): Promise<void> {
