+++ packages/ballerina-visualizer/src/views/BI/AIChatAgent/UseAgentToolForm.tsx
@@ -18,13 +18,13 @@
 
 import { useEffect, useRef, useState } from "react";
 import styled from "@emotion/styled";
-import { buildAgentCallToolNode, FlowNode } from "@wso2/ballerina-core";
+import { FlowNode } from "@wso2/ballerina-core";
 import { FormField, FormValues } from "@wso2/ballerina-side-panel";
 import { Icon } from "@wso2/ui-toolkit";
 import { useRpcContext } from "@wso2/ballerina-rpc-client";
 import ArtifactForm from "../Forms/ArtifactForm";
 import { RelativeLoader } from "../../../components/RelativeLoader";
-import { addToolToAgentNode } from "./utils";
+import { addToolToAgentNode, buildAgentCallToolNode } from "./utils";
 
 const LoaderContainer = styled.div`
     display: flex;
