+++ packages/ballerina-visualizer/src/views/BI/AIChatAgent/utils.ts
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-import { AvailableNode, CodeData, ConfigVariable, EVENT_TYPE, FlowNode, LinePosition, LineRange, NodeKind, NodePosition, ProjectStructureArtifactResponse, Property, SearchNodesQueryParams } from "@wso2/ballerina-core";
+import { AgentToolData, AvailableNode, CodeData, ConfigVariable, EVENT_TYPE, FlowNode, LinePosition, LineRange, NodeKind, NodePosition, ProjectStructureArtifactResponse, Property, SearchNodesQueryParams, ToolParameters } from "@wso2/ballerina-core";
 import { BallerinaRpcClient } from "@wso2/ballerina-rpc-client";
 import { cloneDeep } from "lodash";
 import { URI, Utils } from "vscode-uri";
@@ -533,6 +533,45 @@ export const addToolToAgentNode = async (agentNode: FlowNode, toolName: string) 
     updatedAgentNode.codedata.isNew = false;
     return updatedAgentNode;
 };
+
+export function buildAgentToolNode(wrappedNode: FlowNode, toolName: string, description: string, connection: string,
+    toolParameters?: ToolParameters): FlowNode {
+    const auth = wrappedNode.codedata.data?.auth;
+    const data: AgentToolData = {
+        node: wrappedNode,
+        connection,
+        description,
+        ...(typeof auth === "string" ? { auth } : {}),
+    };
+    return createAgentToolNode(toolName, data, toolParameters ? { parameters: toolParameters } : {});
+}
+
+export function buildAgentCallToolNode(toolName: string, agentVarName: string, includeContext: boolean,
+    description: string): FlowNode {
+    const data: AgentToolData = { toolKind: "AGENT_CALL", agentVarName, includeContext, description };
+    return createAgentToolNode(toolName, data);
+}
+
+function createAgentToolNode(toolName: string, data: AgentToolData, extraProperties: FlowNode["properties"] = {}): FlowNode {
+    return {
+        id: "0",
+        metadata: { label: "Agent Tool", description: "" },
+        codedata: { node: "AGENT_TOOL", isNew: true, data },
+        properties: {
+            functionName: {
+                metadata: { label: "Name", description: "Name of the tool" },
+                valueType: "IDENTIFIER",
+                value: toolName,
+                optional: false,
+                editable: true,
+                advanced: false,
+            } as Property,
+            ...extraProperties,
+        },
+        branches: [],
+        returning: false,
+    };
+}
 
 export interface McpServerConfig {
     name: string;
